WIP: Add some descriptions to errors
I know I'm editing a generated file. This was only to find out what is
causing the error.

The problem was this: I connected to a DigitalOcean Space with the wrong
key and all I got was empty error messages. This is because in the
`NoSuchKey` case, the message fiels is empty.

As a workaround I added the enum variant name in front of the error
message. If this makes sense it can probably be easily done in the generator
script.